third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-setLocalDescription-parameterless.https-expected.txt

This is a testharness.js-based test.
[FAIL] Parameterless SLD() uses [[LastCreatedOffer]] if it is still valid
  assert_true: offerer.pendingLocalDescription.sdp == offer.sdp expected true got false
[FAIL] Parameterless SLD() uses [[LastCreatedAnswer]] if it is still valid
  assert_true: answerer.currentLocalDescription.sdp == answer.sdp expected true got false
[FAIL] RTCSessionDescription constructed without type throws TypeError
  assert_equals: expected "TypeError" but got "ReferenceError"
Harness: the test ran to completion.
